+++ packages/moengage_flutter/moengage_flutter_android/CHANGELOG.md
@@ -1,5 +1,11 @@
 # MoEngage Flutter Android Plugin
 
+# Release Date
+
+## Release Version
+
+- [minor] Added support Logout Complete listener
+
 # 07-05-2026
 
 ## 4.8.0